AWS Provider test
